etlgcp/requirements.txt
google-cloud-bigquery==1.24.0
google-cloud-logging==1.15.0
google-cloud-pubsub==1.4.3
google-cloud-storage==1.28.1
